Don't try to delete nonexistent files
This will give warnings otherwise on skipping platforms

+++ t/find2perl.t
@@ -222,8 +222,8 @@ for my $test (@testcases) {
 }
 
 END {
-    remove_tree($tmpdir);
-    remove_tree($script);
+    remove_tree($tmpdir) if defined $tmpdir;
+    remove_tree($script) if defined $script;
 }
 
 sub runperl {
